feat: Add Sorting, Stack, and Graph visualizer components and integrate them into the DSA visualization page.
