+++ src/pages/DsaVisualization.jsx
@@ -2,23 +2,23 @@
 import React, { useState, useEffect, useRef, useMemo } from 'react';
 import { motion, AnimatePresence } from 'framer-motion';
 import { Helmet } from 'react-helmet';
-import { 
-  Play, Pause, RotateCcw, Settings, ChevronRight, ChevronLeft, 
-  LayoutGrid, List, BarChart2, Activity, Moon, Sun, 
-  GitCompare, SplitSquareHorizontal, ArrowRight, Network, 
-  PlusCircle, Move, MousePointer2, Trash2, StepForward, 
-  StepBack, Binary, BoxSelect, Layers, Search, Github
+import {
+    Play, Pause, RotateCcw, Settings, ChevronRight, ChevronLeft,
+    LayoutGrid, List, BarChart2, Activity, Moon, Sun,
+    GitCompare, SplitSquareHorizontal, ArrowRight, Network,
+    PlusCircle, Move, MousePointer2, Trash2, StepForward,
+    StepBack, Binary, BoxSelect, Layers, Search, Github
 } from 'lucide-react';
 import { Button } from '@/components/ui/button';
 import { Slider } from '@/components/ui/slider';
 import { Switch } from '@/components/ui/switch';
 import { Input } from '@/components/ui/input';
 import {
-  Select,
-  SelectContent,
-  SelectItem,
-  SelectTrigger,
-  SelectValue,
+    Select,
+    SelectContent,
+    SelectItem,
+    SelectTrigger,
+    SelectValue,
 } from "@/components/ui/select";
 import { Card, CardContent } from '@/components/ui/card';
 import { Tabs, TabsList, TabsTrigger, TabsContent } from "@/components/ui/tabs";
@@ -31,368 +31,10 @@ import LinkedListVisualizer from '@/components/dsa/LinkedListVisualizer';
 import TreeVisualizer from '@/components/dsa/TreeVisualizer';
 import HeapVisualizer from '@/components/dsa/HeapVisualizer';
 import DPVisualizer from '@/components/dsa/DPVisualizer';
+import SortingVisualizer from '@/components/dsa/SortingVisualizer';
+import GraphVisualizer from '@/components/dsa/GraphVisualizer';
+import StackVisualizer from '@/components/dsa/StackVisualizer';
 import ContributorsSection from '@/components/dsa/ContributorsSection';
-
-// --- Existing Sorting/Graph Components ---
-
-const SortingVisualizer = ({ array, algorithmName, isPlaying, speed, onFinished, className, searchTarget }) => {
-    const [displayArray, setDisplayArray] = useState([...array]);
-    const timeoutsRef = useRef([]);
-    
-    const showLabels = array.length <= 35;
-
-    useEffect(() => {
-        setDisplayArray([...array]);
-        clearAllTimeouts();
-    }, [array]);
-
-    useEffect(() => {
-        if (isPlaying) {
-            runAlgorithm();
-        } else {
-            clearAllTimeouts();
-        }
-        return () => clearAllTimeouts();
-    }, [isPlaying, algorithmName]);
-
-    const clearAllTimeouts = () => {
-        timeoutsRef.current.forEach(clearTimeout);
-        timeoutsRef.current = [];
-    };
-
-    const runAlgorithm = () => {
-        clearAllTimeouts();
-        let steps = [];
-        const algoFunc = algorithms[`generate${algorithmName}Steps`];
-        if (!algoFunc) return;
-        
-        // Pass searchTarget if it's a search algorithm
-        if (algorithmName.includes('Search')) {
-            steps = algoFunc(array, searchTarget);
-        } else {
-            steps = algoFunc(array);
-        }
-        
-        animate(steps);
-    };
-
-    const animate = (steps) => {
-        const delay = Math.max(1, 100 - speed);
-        steps.forEach((step, index) => {
-            const timeoutId = setTimeout(() => {
-                setDisplayArray(prev => {
-                    // Use a new array reference to trigger re-renders
-                    const newArr = [...prev];
-                    if (step.type === 'swap') {
-                        newArr[step.indices[0]] = step.values[0];
-                        newArr[step.indices[1]] = step.values[1];
-                    } else if (step.type === 'overwrite') {
-                        newArr[step.indices[0]] = step.value;
-                    }
-                    return newArr;
-                });
-
-                const bars = document.getElementsByClassName(`bar-${className}`);
-                if(bars.length > 0) {
-                    if (step.type === 'compare') {
-                         step.indices.forEach(idx => { if(bars[idx]) bars[idx].style.backgroundColor = '#ef4444' });
-                    } else if (step.type === 'swap' || step.type === 'overwrite') {
-                         step.indices.forEach(idx => { if(bars[idx]) bars[idx].style.backgroundColor = '#eab308' });
-                    } else if (step.type === 'sorted' || step.type === 'found') {
-                         step.indices.forEach(idx => { if(bars[idx]) bars[idx].style.backgroundColor = '#10b981' });
-                    } else if (step.type === 'revert') {
-                        step.indices.forEach(idx => { if(bars[idx]) bars[idx].style.backgroundColor = '' }); 
-                    } else if (step.type === 'discard') {
-                         // Gray out discarded parts of the array for Binary Search
-                         if (step.range) {
-                             for(let i=step.range[0]; i<=step.range[1]; i++) {
-                                 if(bars[i]) bars[i].style.opacity = '0.2';
-                             }
-                         }
-                    }
-                }
-
-                if (index === steps.length - 1) {
-                    if (onFinished) onFinished();
-                }
-            }, index * delay);
-            timeoutsRef.current.push(timeoutId);
-        });
-    };
-
-    return (
-        <div className={cn("flex items-end justify-center h-64 w-full p-4 bg-slate-900/50 rounded-xl border border-slate-800 select-none", className, showLabels ? "gap-1" : "gap-[1px]")}>
-            {displayArray.map((val, idx) => (
-                <div key={idx} className="relative flex flex-col justify-end items-center w-full h-full group">
-                    {showLabels && (
-                        <span 
-                            className="absolute text-[10px] font-bold text-slate-500 dark:text-slate-400 transition-all duration-75 text-center w-full pointer-events-none z-10"
-                            style={{ 
-                                bottom: `${(val / Math.max(...array)) * 100}%`, 
-                                marginBottom: '2px',
-                                textShadow: '0 1px 2px rgba(0,0,0,0.1)' 
-                            }}
-                        >
-                            {val}
-                        </span>
-                    )}
-                    <div
-                        className={`bar-${className} w-full rounded-t-md transition-all duration-75 ease-linear hover:opacity-80`}
-                        style={{
-                            height: `${(val / Math.max(...array)) * 100}%`,
-                            backgroundColor: 'rgba(148, 163, 184, 0.5)' 
-                        }}
-                    />
-                </div>
-            ))}
-        </div>
-    );
-};
-
-const GraphVisualizer = ({ algorithmName, isPlaying, setIsPlaying, speed, onFinished, darkMode }) => {
-    const [nodes, setNodes] = useState([]);
-    const [edges, setEdges] = useState([]);
-    const [isDirected, setIsDirected] = useState(false);
-    const [isWeighted, setIsWeighted] = useState(false);
-    const [editMode, setEditMode] = useState('move'); 
-    const [history, setHistory] = useState([]);
-    const [currentStep, setCurrentStep] = useState(-1);
-    const [startNode, setStartNode] = useState(0);
-    const [endNode, setEndNode] = useState(null);
-    const [draggingNode, setDraggingNode] = useState(null);
-    const [edgeStartNode, setEdgeStartNode] = useState(null);
-    const containerRef = useRef(null);
-    const animationRef = useRef(null);
-
-    const visualState = useMemo(() => {
-        const visited = new Set();
-        const path = new Set();
-        const dists = {};
-        let activeLink = null;
-
-        if (currentStep >= 0 && currentStep < history.length) {
-            for (let i = 0; i <= currentStep; i++) {
-                const step = history[i];
-                if (step.type === 'visit' || step.type === 'visit-node') {
-                    visited.add(step.node);
-                } else if (step.type === 'path') {
-                    step.path.forEach(n => path.add(n));
-                } else if (step.type === 'update-dist') {
-                    dists[step.node] = step.dist;
-                }
-                if (i === currentStep) {
-                    if (step.type === 'traverse' || step.type === 'check-edge') {
-                        activeLink = { from: step.from, to: step.to };
-                    }
-                }
-            }
-        }
-        return { visited, path, dists, activeLink };
-    }, [currentStep, history]);
-
-    useEffect(() => { resetGraph(); }, []);
-
-    useEffect(() => {
-        if (!isPlaying && nodes.length > 0) {
-           calculateAlgorithmSteps();
-        }
-    }, [nodes, edges, startNode, endNode, algorithmName, isDirected]);
-
-    const calculateAlgorithmSteps = () => {
-        const numNodes = Math.max(...nodes.map(n => n.id), -1) + 1;
-        const adj = Array.from({ length: numNodes }, () => []);
-        edges.forEach(edge => {
-            if(nodes.find(n=>n.id === edge.from) && nodes.find(n=>n.id === edge.to)) {
-                 adj[edge.from].push({ to: edge.to, weight: edge.weight });
-                 if (!isDirected) adj[edge.to].push({ to: edge.from, weight: edge.weight });
-            }
-        });
-        if (!nodes.find(n => n.id === startNode)) return;
-        let actualEnd = endNode;
-        if (endNode === null && nodes.length > 0) actualEnd = nodes[nodes.length-1].id;
-
-        const algoFunc = algorithms[`generate${algorithmName}Steps`];
-        if (algoFunc) {
-            let steps = [];
-            if (algorithmName === 'AStar') {
-                 const nodeObjMap = nodes.reduce((acc, n) => ({...acc, [n.id]: n}), {});
-                 steps = algoFunc(numNodes, adj, startNode, actualEnd, nodeObjMap);
-            } else {
-                 steps = algoFunc(numNodes, adj, startNode, actualEnd);
-            }
-            setHistory(steps);
-        }
-    };
-
-    useEffect(() => {
-        if (isPlaying) {
-            const delay = Math.max(50, 1000 - (speed * 9));
-            animationRef.current = setInterval(() => {
-                setCurrentStep(prev => {
-                    if (prev < history.length - 1) return prev + 1;
-                    setIsPlaying(false);
-                    if (onFinished) onFinished();
-                    return prev;
-                });
-            }, delay);
-        } else {
-            clearInterval(animationRef.current);
-        }
-        return () => clearInterval(animationRef.current);
-    }, [isPlaying, history.length, speed]);
-
-    const resetGraph = () => {
-        setIsPlaying(false);
-        setCurrentStep(-1);
-        const width = containerRef.current ? containerRef.current.clientWidth : 800;
-        const height = 400;
-        const numNodes = 10;
-        const newNodes = [];
-        for (let i = 0; i < numNodes; i++) newNodes.push({ id: i, x: Math.random() * (width - 100) + 50, y: Math.random() * (height - 100) + 50 });
-        const newEdges = [];
-        for(let i = 1; i < numNodes; i++) {
-            const target = Math.floor(Math.random() * i);
-            const weight = Math.floor(Math.random() * 20) + 1;
-            newEdges.push({ from: i, to: target, weight });
-        }
-        for(let i=0; i<numNodes/2; i++) {
-            const u = Math.floor(Math.random()*numNodes);
-            const v = Math.floor(Math.random()*numNodes);
-            if(u!==v && !newEdges.find(e=>(e.from===u&&e.to===v) || (e.from===v&&e.to===u))) {
-                 newEdges.push({ from: u, to: v, weight: Math.floor(Math.random()*20)+1 });
-            }
-        }
-        setNodes(newNodes);
-        setEdges(newEdges);
-        setStartNode(0);
-        setEndNode(numNodes-1);
-    };
-
-    const handleMouseDown = (e, nodeId) => {
-        if (isPlaying) return;
-        e.stopPropagation();
-        if (editMode === 'move' && nodeId !== undefined) setDraggingNode(nodeId);
-        else if (editMode === 'add-edge' && nodeId !== undefined) {
-            if (edgeStartNode === null) setEdgeStartNode(nodeId);
-            else {
-                if (edgeStartNode !== nodeId) {
-                    const weight = isWeighted ? Math.floor(Math.random() * 20) + 1 : 1;
-                    const exists = edges.some(edge => (edge.from === edgeStartNode && edge.to === nodeId) || (!isDirected && edge.from === nodeId && edge.to === edgeStartNode));
-                    if (!exists) setEdges(prev => [...prev, { from: edgeStartNode, to: nodeId, weight }]);
-                }
-                setEdgeStartNode(null);
-            }
-        } else if (editMode === 'delete' && nodeId !== undefined) {
-            setNodes(prev => prev.filter(n => n.id !== nodeId));
-            setEdges(prev => prev.filter(e => e.from !== nodeId && e.to !== nodeId));
-            if(startNode === nodeId) setStartNode(nodes.length > 0 ? nodes[0].id : null);
-        }
-    };
-
-    const handleCanvasClick = (e) => {
-        if (isPlaying) return;
-        if (editMode === 'add-node') {
-            const rect = containerRef.current.getBoundingClientRect();
-            const x = e.clientX - rect.left;
-            const y = e.clientY - rect.top;
-            const newId = nodes.length > 0 ? Math.max(...nodes.map(n=>n.id)) + 1 : 0;
-            setNodes(prev => [...prev, { id: newId, x, y }]);
-        }
-    };
-
-    const handleMouseMove = (e) => {
-        if (draggingNode === null || !containerRef.current) return;
-        const rect = containerRef.current.getBoundingClientRect();
-        const x = e.clientX - rect.left;
-        const y = e.clientY - rect.top;
-        setNodes(prev => prev.map(n => n.id === draggingNode ? { ...n, x, y } : n));
-    };
-
-    const handleMouseUp = () => setDraggingNode(null);
-    
-    const handleNodeContext = (e, nodeId) => {
-        e.preventDefault();
-        if(editMode === 'move') { setStartNode(nodeId); setCurrentStep(-1); }
-    };
-
-    return (
-        <div className="flex flex-col gap-4">
-            <div className={cn("flex flex-wrap gap-4 p-4 rounded-lg border", darkMode ? "bg-slate-900 border-slate-800" : "bg-white border-slate-200")}>
-                {/* Graph Toolbar Buttons: Added explicit colors for visibility */}
-                <div className="flex items-center gap-1 bg-slate-100 dark:bg-slate-800 p-1 rounded-md">
-                    <Button 
-                        size="sm" variant={editMode === 'move' ? "default" : "ghost"} 
-                        className={cn(editMode !== 'move' && "text-slate-700 dark:text-slate-200")}
-                        onClick={() => setEditMode('move')} title="Move Nodes"
-                    >
-                        <Move className="h-4 w-4" />
-                    </Button>
-                    <Button 
-                        size="sm" variant={editMode === 'add-node' ? "default" : "ghost"} 
-                        className={cn(editMode !== 'add-node' && "text-slate-700 dark:text-slate-200")}
-                        onClick={() => setEditMode('add-node')} title="Add Node"
-                    >
-                        <PlusCircle className="h-4 w-4" />
-                    </Button>
-                    <Button 
-                        size="sm" variant={editMode === 'add-edge' ? "default" : "ghost"} 
-                        className={cn(editMode !== 'add-edge' && "text-slate-700 dark:text-slate-200")}
-                        onClick={() => setEditMode('add-edge')} title="Add Edge"
-                    >
-                        <Network className="h-4 w-4" />
-                    </Button>
-                    <Button 
-                        size="sm" variant={editMode === 'delete' ? "destructive" : "ghost"} 
-                        className={cn(editMode !== 'delete' && "text-slate-700 dark:text-slate-200")}
-                        onClick={() => setEditMode('delete')} title="Delete Node"
-                    >
-                        <Trash2 className="h-4 w-4" />
-                    </Button>
-                </div>
-                <div className="w-px h-8 bg-slate-200 dark:bg-slate-700 mx-1" />
-                <div className="flex items-center gap-4 text-xs font-medium">
-                    <div className="flex items-center gap-2"><Switch checked={isDirected} onCheckedChange={setIsDirected} id="directed-mode" /><label htmlFor="directed-mode">Directed</label></div>
-                    <div className="flex items-center gap-2"><Switch checked={isWeighted} onCheckedChange={setIsWeighted} id="weighted-mode" /><label htmlFor="weighted-mode">Weighted</label></div>
-                </div>
-                <div className="w-px h-8 bg-slate-200 dark:bg-slate-700 mx-1 ml-auto" />
-                <div className="flex items-center gap-2">
-                     <Button size="icon" variant="outline" disabled={currentStep <= -1} onClick={() => { setIsPlaying(false); setCurrentStep(prev => Math.max(-1, prev - 1)); }}><StepBack className="h-4 w-4" /></Button>
-                     <Button size="icon" variant={isPlaying ? "destructive" : "default"} onClick={() => setIsPlaying(!isPlaying)}>{isPlaying ? <Pause className="h-4 w-4" /> : <Play className="h-4 w-4" />}</Button>
-                     <Button size="icon" variant="outline" disabled={currentStep >= history.length - 1} onClick={() => { setIsPlaying(false); setCurrentStep(prev => Math.min(history.length - 1, prev + 1)); }}><StepForward className="h-4 w-4" /></Button>
-                </div>
-            </div>
-
-            <div ref={containerRef} className={cn("relative w-full h-[450px] border rounded-xl overflow-hidden select-none", darkMode ? "bg-slate-950 border-slate-800" : "bg-slate-50 border-slate-200", editMode === 'add-node' && "cursor-crosshair")} onClick={handleCanvasClick} onMouseMove={handleMouseMove} onMouseUp={handleMouseUp} onMouseLeave={handleMouseUp}>
-                <div className="absolute top-4 left-4 z-10 pointer-events-none">
-                    <div className="bg-black/60 backdrop-blur text-white text-xs px-3 py-2 rounded-lg space-y-1">
-                        <p><span className="text-emerald-400 font-bold">Start:</span> {startNode}</p>
-                        <p><span className="text-red-400 font-bold">End:</span> {endNode ?? 'Last'}</p>
-                        <p className="opacity-80">{editMode === 'move' ? "Drag nodes. Right-click to set Start." : "Click to edit."}</p>
-                    </div>
-                </div>
-                <svg width="0" height="0"><defs><marker id="arrowhead" markerWidth="10" markerHeight="7" refX="28" refY="3.5" orient="auto"><polygon points="0 0, 10 3.5, 0 7" fill={darkMode ? "#64748b" : "#94a3b8"} /></marker><marker id="arrowhead-active" markerWidth="10" markerHeight="7" refX="28" refY="3.5" orient="auto"><polygon points="0 0, 10 3.5, 0 7" fill="#eab308" /></marker></defs></svg>
-                <svg width="100%" height="100%" className="pointer-events-none">
-                    {edges.map((edge, idx) => {
-                        const start = nodes.find(n => n.id === edge.from);
-                        const end = nodes.find(n => n.id === edge.to);
-                        if (!start || !end) return null;
-                        const isActive = visualState.activeLink && ((visualState.activeLink.from === edge.from && visualState.activeLink.to === edge.to) || (!isDirected && visualState.activeLink.from === edge.to && visualState.activeLink.to === edge.from));
-                        return ( <g key={`edge-${idx}`}><line x1={start.x} y1={start.y} x2={end.x} y2={end.y} stroke={isActive ? '#eab308' : (darkMode ? '#334155' : '#cbd5e1')} strokeWidth={isActive ? 3 : 2} markerEnd={isDirected ? (isActive ? "url(#arrowhead-active)" : "url(#arrowhead)") : undefined} className="transition-colors duration-200" />{isWeighted && (<g transform={`translate(${(start.x+end.x)/2}, ${(start.y+end.y)/2})`}><rect x="-8" y="-8" width="16" height="16" rx="4" fill={darkMode ? "#0f172a" : "#ffffff"} /><text dy="4" textAnchor="middle" fill={darkMode ? '#94a3b8' : '#475569'} fontSize="10" fontWeight="bold">{edge.weight}</text></g>)}</g>);
-                    })}
-                     {editMode === 'add-edge' && edgeStartNode !== null && (() => { const start = nodes.find(n => n.id === edgeStartNode); if(start) return <circle cx={start.x} cy={start.y} r={22} fill="none" stroke="#eab308" strokeWidth="2" strokeDasharray="4 4" className="animate-spin-slow" />; return null; })()}
-                    {nodes.map((node) => {
-                        const isStart = node.id === startNode;
-                        const isEnd = node.id === endNode;
-                        const isVisited = visualState.visited.has(node.id);
-                        const isPath = visualState.path.has(node.id);
-                        const dist = visualState.dists[node.id];
-                        return (<g key={node.id} transform={`translate(${node.x},${node.y})`} className="pointer-events-auto cursor-pointer transition-transform duration-200 hover:scale-110" onMouseDown={(e) => handleMouseDown(e, node.id)} onClick={() => editMode === 'move' && setEndNode(node.id)} onContextMenu={(e) => handleNodeContext(e, node.id)}><circle r={18} fill={isStart ? '#10b981' : isEnd ? '#ef4444' : isPath ? '#8b5cf6' : isVisited ? '#3b82f6' : (darkMode ? '#1e293b' : '#ffffff')} stroke={isStart || isEnd || isPath || isVisited ? 'white' : (darkMode ? '#475569' : '#94a3b8')} strokeWidth={2} className="transition-colors duration-300" /><text dy=".3em" textAnchor="middle" fill={isStart || isEnd || isPath || isVisited ? 'white' : (darkMode ? '#e2e8f0' : '#1e293b')} fontSize="12" fontWeight="bold" className="pointer-events-none">{node.id}</text>{(algorithmName === 'Dijkstra' || algorithmName === 'AStar') && dist !== undefined && (<text dy="-1.8em" textAnchor="middle" fill={darkMode ? '#cbd5e1' : '#475569'} fontSize="10" className="font-mono">{dist === Infinity ? '∞' : dist}</text>)}</g>);
-                    })}
-                </svg>
-            </div>
-        </div>
-    );
-};
 
 // --- Main Page ---
 
@@ -401,14 +43,14 @@ const DsaVisualization = () => {
     const [activeTab, setActiveTab] = useState('sorting');
     const [algorithm, setAlgorithm] = useState('BubbleSort');
     const [secondAlgorithm, setSecondAlgorithm] = useState('QuickSort');
-    const [arraySize, setArraySize] = useState(20); 
+    const [arraySize, setArraySize] = useState(20);
     const [speed, setSpeed] = useState(50);
     const [isPlaying, setIsPlaying] = useState(false);
     const [isComparisonMode, setIsComparisonMode] = useState(false);
     const [darkMode, setDarkMode] = useState(true);
     const [customInput, setCustomInput] = useState('');
     const [searchTarget, setSearchTarget] = useState(42); // Default target
-    
+
     const [array, setArray] = useState([]);
 
     useEffect(() => {
@@ -423,7 +65,7 @@ const DsaVisualization = () => {
         // To keep it simple, we generate random. 
         // If user selects BinarySearch, we might want to auto-sort or let them visualize fail/sort.
         const newArr = Array.from({ length: arraySize }, () => Math.floor(Math.random() * 100) + 5);
-        
+
         // If searching, ensure the target is present randomly 50% of the time for fun, or just random.
         // Let's just keep random.
         setArray(newArr);
@@ -433,11 +75,11 @@ const DsaVisualization = () => {
             bar.style.opacity = '1';
         });
     };
-    
+
     // Auto-sort if switching to Binary Search to avoid confusion
     useEffect(() => {
         if (algorithm === 'BinarySearch') {
-            setArray(prev => [...prev].sort((a,b) => a-b));
+            setArray(prev => [...prev].sort((a, b) => a - b));
         }
     }, [algorithm]);
 
@@ -456,7 +98,8 @@ const DsaVisualization = () => {
         linkedlist: [],
         trees: [],
         heaps: [],
-        dp: []
+        dp: [],
+        stack: []
     };
 
     return (
@@ -499,8 +142,8 @@ const DsaVisualization = () => {
                             </p>
                         </div>
                     </div>
-                    <Button 
-                        asChild 
+                    <Button
+                        asChild
                         className={cn(
                             "w-full md:w-auto px-6 py-2 rounded-md font-semibold transition-all duration-300",
                             darkMode ? "bg-cyan-500 hover:bg-cyan-600 text-white" : "bg-blue-600 hover:bg-blue-700 text-white"
@@ -518,15 +161,15 @@ const DsaVisualization = () => {
                             <div className="space-y-4">
                                 <h3 className="font-semibold text-sm uppercase tracking-wider text-slate-500">Category</h3>
                                 <div className="grid grid-cols-2 gap-2">
-                                    {['sorting', 'searching', 'graphs', 'linkedlist', 'trees', 'heaps', 'dp'].map(cat => (
+                                    {['sorting', 'searching', 'graphs', 'linkedlist', 'trees', 'heaps', 'dp', 'stack'].map(cat => (
                                         <Button
                                             key={cat}
                                             variant={activeTab === cat ? 'default' : 'outline'}
-                                            onClick={() => { 
-                                                setActiveTab(cat); 
-                                                setIsPlaying(false); 
+                                            onClick={() => {
+                                                setActiveTab(cat);
+                                                setIsPlaying(false);
                                                 setIsComparisonMode(false);
-                                                if(algoOptions[cat]?.length > 0) setAlgorithm(algoOptions[cat][0]);
+                                                if (algoOptions[cat]?.length > 0) setAlgorithm(algoOptions[cat][0]);
                                             }}
                                             className="w-full text-xs capitalize"
                                         >
@@ -546,9 +189,9 @@ const DsaVisualization = () => {
                                         {/* Fix: Explicit text colors for dropdown items to ensure contrast */}
                                         <SelectContent className={cn(darkMode ? "bg-slate-900 border-slate-800" : "bg-white border-slate-200")}>
                                             {algoOptions[activeTab]?.map(algo => (
-                                                <SelectItem 
-                                                    key={algo} 
-                                                    value={algo} 
+                                                <SelectItem
+                                                    key={algo}
+                                                    value={algo}
                                                     className={cn(
                                                         "cursor-pointer",
                                                         darkMode ? "text-slate-100 focus:text-white focus:bg-slate-800" : "text-slate-900 focus:text-black focus:bg-slate-100"
@@ -570,9 +213,9 @@ const DsaVisualization = () => {
                                             <SelectTrigger className="mt-2"><SelectValue placeholder="Select 2nd Algorithm" /></SelectTrigger>
                                             <SelectContent className={cn(darkMode ? "bg-slate-900 border-slate-800" : "bg-white border-slate-200")}>
                                                 {algoOptions[activeTab]?.map(algo => (
-                                                    <SelectItem 
-                                                        key={algo} 
-                                                        value={algo} 
+                                                    <SelectItem
+                                                        key={algo}
+                                                        value={algo}
                                                         className={cn(
                                                             "cursor-pointer",
                                                             darkMode ? "text-slate-100 focus:text-white focus:bg-slate-800" : "text-slate-900 focus:text-black focus:bg-slate-100"
@@ -591,7 +234,7 @@ const DsaVisualization = () => {
                                 <div className="space-y-6">
                                     <div className="space-y-2">
                                         <div className="flex justify-between text-sm"><span>Size</span><span className="text-slate-500">{arraySize}</span></div>
-                                        <Slider value={[arraySize]} max={100} min={10} step={5} onValueChange={(val) => setArraySize(val[0])} disabled={isPlaying}/>
+                                        <Slider value={[arraySize]} max={100} min={10} step={5} onValueChange={(val) => setArraySize(val[0])} disabled={isPlaying} />
                                     </div>
                                     <div className="space-y-2">
                                         <div className="flex justify-between text-sm"><span>Speed</span><span className="text-slate-500">{speed}%</span></div>
@@ -603,12 +246,12 @@ const DsaVisualization = () => {
                             {(activeTab === 'sorting' || activeTab === 'searching') && (
                                 <div className="pt-4 space-y-3">
                                     <Button className={cn("w-full font-bold", isPlaying ? "bg-red-500 hover:bg-red-600" : "bg-cyan-500 hover:bg-cyan-600")} onClick={handlePlay}>
-                                        {isPlaying ? <><Pause className="mr-2 h-4 w-4"/> Pause</> : <><Play className="mr-2 h-4 w-4"/> Start</>}
+                                        {isPlaying ? <><Pause className="mr-2 h-4 w-4" /> Pause</> : <><Play className="mr-2 h-4 w-4" /> Start</>}
                                     </Button>
                                     <Button variant="outline" className="w-full" onClick={resetArray} disabled={isPlaying}><RotateCcw className="mr-2 h-4 w-4" /> Reset / Randomize</Button>
                                 </div>
                             )}
-                            
+
                             {/* Fix: Added Search Target Input */}
                             {activeTab === 'searching' && (
                                 <div className="space-y-2 pt-4 border-t border-slate-200 dark:border-slate-800">
@@ -616,11 +259,11 @@ const DsaVisualization = () => {
                                         <Search className="w-3 h-3" /> Search Target
                                     </label>
                                     <div className="flex gap-2">
-                                        <Input 
-                                            type="number" 
-                                            placeholder="Target Value" 
-                                            value={searchTarget} 
-                                            onChange={(e) => setSearchTarget(parseInt(e.target.value) || 0)} 
+                                        <Input
+                                            type="number"
+                                            placeholder="Target Value"
+                                            value={searchTarget}
+                                            onChange={(e) => setSearchTarget(parseInt(e.target.value) || 0)}
                                             className="h-9 text-sm"
                                         />
                                     </div>
@@ -650,11 +293,11 @@ const DsaVisualization = () => {
                                         </div>
                                         <SortingVisualizer array={array} algorithmName={algorithm} isPlaying={isPlaying} speed={speed} className="primary" searchTarget={searchTarget} />
                                     </CardContent>
-                                공급                                </Card>
+                                    공급                                </Card>
                                 {isComparisonMode && (
                                     <Card className={cn("overflow-hidden border-0 shadow-2xl", darkMode ? "bg-slate-900" : "bg-white")}>
                                         <CardContent className="p-6">
-                                             <div className="flex justify-between items-center mb-4">
+                                            <div className="flex justify-between items-center mb-4">
                                                 <h3 className="font-bold text-xl text-purple-400">{secondAlgorithm.replace(/([A-Z])/g, ' $1').trim()}</h3>
                                             </div>
                                             <SortingVisualizer array={array} algorithmName={secondAlgorithm} isPlaying={isPlaying} speed={speed} className="secondary" searchTarget={searchTarget} />
@@ -672,7 +315,7 @@ const DsaVisualization = () => {
                                 </CardContent>
                             </Card>
                         ) : activeTab === 'linkedlist' ? (
-                             <Card className={cn("overflow-hidden border-0 shadow-2xl", darkMode ? "bg-slate-900" : "bg-white")}>
+                            <Card className={cn("overflow-hidden border-0 shadow-2xl", darkMode ? "bg-slate-900" : "bg-white")}>
                                 <CardContent className="p-6">
                                     <div className="flex justify-between items-center mb-4">
                                         <h3 className="font-bold text-xl text-emerald-400 flex items-center gap-2"><List className="h-5 w-5" /> Linked List Visualizer</h3>
@@ -690,7 +333,7 @@ const DsaVisualization = () => {
                                 </CardContent>
                             </Card>
                         ) : activeTab === 'heaps' ? (
-                             <Card className={cn("overflow-hidden border-0 shadow-2xl", darkMode ? "bg-slate-900" : "bg-white")}>
+                            <Card className={cn("overflow-hidden border-0 shadow-2xl", darkMode ? "bg-slate-900" : "bg-white")}>
                                 <CardContent className="p-6">
                                     <div className="flex justify-between items-center mb-4">
                                         <h3 className="font-bold text-xl text-yellow-400 flex items-center gap-2"><Layers className="h-5 w-5" /> Heap Visualizer (Max)</h3>
@@ -699,7 +342,7 @@ const DsaVisualization = () => {
                                 </CardContent>
                             </Card>
                         ) : activeTab === 'dp' ? (
-                             <Card className={cn("overflow-hidden border-0 shadow-2xl", darkMode ? "bg-slate-900" : "bg-white")}>
+                            <Card className={cn("overflow-hidden border-0 shadow-2xl", darkMode ? "bg-slate-900" : "bg-white")}>
                                 <CardContent className="p-6">
                                     <div className="flex justify-between items-center mb-4">
                                         <h3 className="font-bold text-xl text-pink-400 flex items-center gap-2"><BoxSelect className="h-5 w-5" /> Dynamic Programming</h3>
@@ -707,8 +350,17 @@ const DsaVisualization = () => {
                                     <DPVisualizer darkMode={darkMode} />
                                 </CardContent>
                             </Card>
+                        ) : activeTab === 'stack' ? (
+                            <Card className={cn("overflow-hidden border-0 shadow-2xl", darkMode ? "bg-slate-900" : "bg-white")}>
+                                <CardContent className="p-6">
+                                    <div className="flex justify-between items-center mb-4">
+                                        <h3 className="font-bold text-xl text-cyan-500 flex items-center gap-2"><Layers className="h-5 w-5" /> Stack Visualizer</h3>
+                                    </div>
+                                    <StackVisualizer darkMode={darkMode} />
+                                </CardContent>
+                            </Card>
                         ) : (
-                             <Card className={cn("h-96 flex flex-col items-center justify-center border-0 shadow-2xl text-center p-8", darkMode ? "bg-slate-900" : "bg-white")}>
+                            <Card className={cn("h-96 flex flex-col items-center justify-center border-0 shadow-2xl text-center p-8", darkMode ? "bg-slate-900" : "bg-white")}>
                                 <h3 className="text-2xl font-bold">Under Construction</h3>
                             </Card>
                         )}
